+++ packages/eyes-webdriverio-4/src/LegacyWrappedElement.js
@@ -5,6 +5,9 @@
  */
 function LegacyAPIElement(EyesWrappedElement) {
   return class EyesWebElement extends EyesWrappedElement {
+    get __EyesWebElement() {
+      return true
+    }
     get element() {
       return this.unwrapped
     }