+++ packages/eyes-webdriverio-5/src/SpecWrappedDriver.js
@@ -207,7 +207,17 @@ async function visit(driver, url) {
   return driver.url(url)
 }
 
-/* -------- TESTING -------- */
+async function getNativeElementLocation(driver, element) {
+  const extended = await driver.$(element)
+  return extended.getLocation()
+}
+
+async function getNativeElementSize(driver, element) {
+  const extended = await driver.$(element)
+  return extended.getSize()
+}
+
+/* -------- FOR TESTING PURPOSES -------- */
 
 /** @return {WDIOAsyncDriver} */
 async function build({
@@ -306,3 +316,5 @@ exports.click = click
 exports.type = type
 exports.waitUntilDisplayed = waitUntilDisplayed
 exports.getElementRect = getElementRect
+exports.getNativeElementLocation = getNativeElementLocation
+exports.getNativeElementSize = getNativeElementSize