+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,8 +2,21 @@
 
 ## Unreleased
 
+
+## 11.5.0 - 2020/7/26
+
+- support non-200 resources ([Trello](https://trello.com/c/J5lBWutP))
+
+## 11.4.1 - 2020/7/24
+
+- fix issue with duplicate copies of the SDK ([Trello](https://trello.com/c/4Gd7uWtS))
+
+## 11.4.0 - 2020/7/24
+
 - updated enum types to be more explicit about values
 - remove type definitions
+- support Appium mobile selectors
+- updated to @applitools/dom-capture@7.2.6 (from 7.2.4)
 
 ## 11.3.9 - 2020/7/22