+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## Unreleased
 
+
+## 2.20.1 - 2020/7/26
+
+- updated to @applitools/eyes-sdk-core@11.5.0 (from 11.4.1)
+- updated to @applitools/visual-grid-client@14.6.0 (from 14.5.15)
+
+## 2.20.0 - 2020/7/24
+
 - remove type definitions
+- updated to @applitools/eyes-sdk-core@11.4.1 (from 11.3.8)
+- updated to @applitools/visual-grid-client@14.5.15 (from 14.5.12)
 
 ## 2.19.2 - 2020/7/19