+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -3,9 +3,22 @@
 ## Unreleased
 
 - added support of devtools protocol
-- updated to @applitools/eyes-sdk-core@11.2.1 (from 11.2.2)
-- updated to @applitools/visual-grid-client@14.5.1 (from 14.5.2)
+
+## 5.19.2 - 2020/7/26
+
+- updated to @applitools/eyes-sdk-core@11.5.0 (from 11.4.1)
+- updated to @applitools/visual-grid-client@14.6.0 (from 14.5.15)
+
+## 5.19.1 - 2020/7/24
+
+- updated to @applitools/eyes-sdk-core@11.4.1 (from 11.4.0)
+- updated to @applitools/visual-grid-client@14.5.15 (from 14.5.14)
+
+## 5.19.0 - 2020/7/24
+
 - remove type definitions
+- updated to @applitools/eyes-sdk-core@11.4.0 (from 11.3.8)
+- updated to @applitools/visual-grid-client@14.5.14 (from 14.5.12)
 
 ## 5.18.4 - 2020/7/18